Separate request notification banner into its own component
